First Commit
Dump of a bunch of files related to my genetic algorithm stuff. Quite
barebones at the moment, and some hacky stuff I'd like to clean up, but
it works okay.

There's some Netbeans specific stuff in here, cause I'm lazy. So watch
out for that if you want to edit it for whatever reason.