Lookup table version of data set 1 works!!!
Woo! Of course, as the lookup table is literally the simplest way to
solve the problem, and completely fitted to data set 1 as a whole, I
probably shouldn't be celebrating. Still, at least my GA still works...

Also, I tweaked the function that tests the real data. It looks a bit
nicer, and now actually adds the total together.