Changed GA stop condition
You can now tell the GA to stop after it's reached Max Fitness, and
maintained that fitness for a certain number of cycles.

My RuleSet version of data1 passes *most* of the time, but on occasion
fails 1% or 2% of the tests. I should try to figure out why.

Irritatingly, my Roulette selection algorithm doesn't seem to be very
good. I should take a look at that as well.