Cracked it?
May have finally cracked dataset 2. It runs for about 6000 generations,
and gets 99 to 100% each time. Hooray!

On another note, decided to give up on the MLP. It doesn't seem to ever
reach the maximum fitness. I left it running for 50000, and it stayed at
about 1300. Not good.

Gonna tackle dataset 3 with a modified ruleset, instead.